committed a few fixes I found in local dir

+++ filesystem/usr/local/sbin/deleteDb.sh
@@ -1,3 +1,13 @@
-/etc/init.d/postgresql stop
-rm -rf /var/www/postgresql
+#!/bin/bash
+
+echo "Clearing Database"
+
+if [ -d "/var/www/postgresql" ]; then
+	echo "Stopping postgresql and deleting database directory"
+	/etc/init.d/postgresql stop
+	rm -rf /var/www/postgresql
+else
+	echo "No database directory, nothing to clean up"
+fi
+
 rm -rf /var/www/milestones/dbSetup
